% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concensus-default-workflow.R
\name{analyze}
\alias{analyze}
\alias{analyze.concensusDataSet}
\alias{analyze.concensusWorkflow}
\alias{analyze.default}
\title{Run concensusGLM analysis}
\usage{
analyze(x, ...)

\method{analyze}{default}(x, ...)

\method{analyze}{concensusDataSet}(x, grouping = c("compound",
  "concentration", "strain"), conditions = c("compound", "concentration"),
  scattering = "strain", ...)

\method{analyze}{concensusWorkflow}(x, grouping = c("compound",
  "concentration", "strain"), conditions = c("compound", "concentration"),
  scattering = "strain", locality = "local", parallel = FALSE, ...)
}
\arguments{
\item{x}{concensusWorkflow or concensusDataSet.}

\item{...}{Other arguments passed to \link{execute}.}

\item{grouping}{Character vector. Combination of data columns to identify unique observations.}

\item{conditions}{Character vector. Combination of data columns to idenitfy unique conditions.}

\item{scattering}{Character. Column identifying analytically independent chunks for parallelization.}
}
\value{
concensusWorkflow or concensusDataSet with a new \code{mean_variance_relationship} and \code{dispersion} attribute.
}
\description{
Do all caluclations on count data to calculate batch effects, dispersions, and final effect sizes and p-values.
}

