% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concensus-statistical-methods.R
\name{callHits}
\alias{callHits}
\alias{callHits.concensusDataSet}
\alias{callHits.concensusWorkflow}
\alias{callHits.default}
\title{Call hits based on p-value and resampled data}
\usage{
callHits(x, ...)

\method{callHits}{default}(x, ...)

\method{callHits}{concensusWorkflow}(x, resampled_x = NULL, ...)

\method{callHits}{concensusDataSet}(x, method = "resampling",
  false_discovery_rate. = 0.05, prevalence = 0.01, ...)
}
\arguments{
\item{x}{concensusWorkflow or concensusDataSet.}

\item{...}{Other arguments.}

\item{method}{Character. Only \code{"resampling"} is supported.}

\item{prevalence}{Numeric. The expected or known prevalence. Used to construct a model dataset from resampled data.}

\item{false_disovery_rate.}{Numeric. The desired false discovery rate. Default \code{0.05}.}
}
\value{
concensusWorkflow or concensusDataSet.
}
\description{
Call hits based on p-values and resampled data.
}
\details{
Resampled data are used to calaulate FDRs at a sliding p-value cutoff.
}

