% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concensus-objects.R
\name{concensusDataSetFromFile}
\alias{concensusDataSetFromFile}
\title{Set up analysis container for ConcensusGLM}
\usage{
concensusDataSetFromFile(data_filename, annotation_filename = NULL,
  output_path = ".", controls = NULL, rename = NULL, test = FALSE,
  checkpoint = FALSE, ...)
}
\arguments{
\item{data_filename}{Character. Path to a table of counts.}

\item{annotation_filename}{Character. Optional. Path to experimental annotations. These will be joined to the input data
and used for batch correction.}

\item{output_path}{Character. Path to directory where you want the analysis output. Default is current working directory.}

\item{controls}{Named list with elements \code{positive} and \code{negative}.
Adds logical columns to data called \code{positive_control} and \code{negative_control}.}

\item{test}{Logical. Run in test mode?}

\item{checkpoint}{Logical. Save intermediate results as checkpoints?}
}
\value{
list of class "concensusDataSet"
}
\description{
Set up analysis container for ConcensusGLM using path to data as input. Performs validity checks on the input.
Optionally annotates input with experimental metadata from \code{annotation_filename}.
}
\seealso{
\link{workflows::newWorkflow}, \link{concensusDataSetFromFile}
}

