% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concensus-objects.R
\name{directoryTree}
\alias{directoryTree}
\title{Set up a directory tree for ConcensusGLM}
\usage{
directoryTree(path = ".")
}
\arguments{
\item{path}{Character. Path to root of directory tree}
}
\value{
list of class "directoryTree"
}
\description{
Set up a directory tree for ConcensusGLM
}

