% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concensus-methods.R
\name{execute.concensusWorkflow}
\alias{execute.concensusWorkflow}
\title{Execute concensusWorkflow}
\usage{
\method{execute}{concensusWorkflow}(x, locality = "local", parallel = FALSE,
  clobber = FALSE, ...)
}
\arguments{
\item{x}{concensusWorkflow.}

\item{...}{Other arguments passed to \code{workflows::execute}.}
}
\value{
Executed concensusWorkflow
}
\description{
Execute concensusWorkflow analysis pipeline.
}
\seealso{
\link{execute}
}

