% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concensus-statistical-methods.R
\name{getBatchEffects}
\alias{getBatchEffects}
\alias{getBatchEffects.concensusDataSet}
\alias{getBatchEffects.concensusWorkflow}
\alias{getBatchEffects.default}
\title{Get experimental batch effects}
\usage{
getBatchEffects(x, ...)

\method{getBatchEffects}{default}(x, ...)

\method{getBatchEffects}{concensusWorkflow}(x, ...)

\method{getBatchEffects}{concensusDataSet}(x, grouping = c("compound",
  "concentration", "strain"), ...)
}
\arguments{
\item{x}{concensusWorkflow or concensusDataSet.}

\item{...}{Other arguments.}
}
\value{
concensusWorkflow or concensusDataSet with a new \code{batch_effect_model} and a new \code{predicted_null_count} column
in the \code{data} attribute.
}
\description{
Estimate experimental batch effects.
}
\details{
Columns in the \code{data} attribute of the \code{concensusDataSet} object which
have fewer than 100 unique values and which are not named according to a list of stopwords are assumed to be experimental
handling annotations, and will be modeled using a Negative Binomial GLM, using dispersion parameters in the \code{dispersion}
attribute.

Only the negative control observations will be used for estimating batch effects.

A new column, \code{predicted_null_count}, is added to \code{data} of \code{concensusDataSet} with a prediction of
number of counts for every observation, assuming only batch effects (and no real effect of interest).

If the model fitting fails, rather than throw off the whole pipeline, a \code{predicted_null_count} value of \code{1} is used.
}
\seealso{
\link{glm}
}

