% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concensus-statistical-methods.R
\name{getFinalDispersions}
\alias{getFinalDispersions}
\alias{getFinalDispersions.concensusDataSet}
\alias{getFinalDispersions.concensusWorkflow}
\alias{getFinalDispersions.default}
\title{Estimate Negative Binomial dispersion paramteter taking into account batch effects}
\usage{
getFinalDispersions(x, ...)

\method{getFinalDispersions}{default}(x, ...)

\method{getFinalDispersions}{concensusWorkflow}(x, ...)

\method{getFinalDispersions}{concensusDataSet}(x, max_rows = 10000, ...)
}
\arguments{
\item{x}{concensusWorkflow or concensusDataSet.}

\item{...}{Other arguments.}

\item{max_rows}{Numeric. Maximum number of observations to use for MLE.}
}
\value{
concensusWorkflow or concensusDataSet with a new \code{small_model_dispersion} and a new
\code{full_model_dispersion} column in the \code{dispersion} attribute.
}
\description{
Estimate Negative Binomial dispersion paramteter taking into account experimental batch effects.
}
\details{
Uses the CR penalized maximum profile likelihood method, holding the \eqn{\mu} of a GLM fixed and finding the
optimal dispersion \eqn{\alpha} using a Newton-type algorithm as implemented in \code{nlm}.

If the \code{predicted_null_count} column is present in the \code{data} attribute of \code{concensusDataSet}, it is added to
the GLM as an \code{offset}. If the batch effects are real, this should raise the likelihood of and shrink the size of
the final dispersion parameter.

This method will find a dispersion value with and without taking into account \code{predicted_null_count}, saving both results to
the columns of the \code{dispersion} attribute of \code{concensusDataSet}.
}
\seealso{
\link{nlm}, \link{glm}
}

