% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concensus-statistical-methods.R
\name{getFinalModel}
\alias{getFinalModel}
\alias{getFinalModel.concensusDataSet}
\alias{getFinalModel.concensusWorkflow}
\alias{getFinalModel.default}
\title{Fit final model}
\usage{
getFinalModel(x, ...)

\method{getFinalModel}{default}(x, ...)

\method{getFinalModel}{concensusWorkflow}(x, ...)

\method{getFinalModel}{concensusDataSet}(x, conditions = c("compound",
  "concentration"), grouping = "strain", ...)
}
\arguments{
\item{x}{concensusWorkflow or concensusDataSet.}

\item{...}{Other arguments.}

\item{conditions}{Character vector. Columns in the \code{data} attribute of \code{concensusDataSet} which together identify groups
of replicates of the same condition.}

\item{grouping}{Character vector. Columns in the \code{data} attribute of \code{concensusDataSet} which together identify
analytically independent chunks of data, e.g. strains.}
}
\value{
concensusWorkflow or concensusDataSet with a new \code{model_parameters} attribute containing effect sizes (LFC) and
p-values.
}
\description{
Fit the final Negative Binomial GLM, taking into account batch effects and final dispersion estimates.
}
\details{
Using \code{predicted_null_count} column in the \code{data} attribute of \code{concensusDataSet} is added to
the GLM as an \code{offset}. A new column, \code{condition_group} is created as a concatenation of the combinations of values
encountered in the columns specified in \code{conditions}, which is used as a catagorical variable.

The \code{condition_group} associated with the negative control is identified, and set as the reference. A Negative Binomial
GLM is fitted using no intercept, \code{predicted_null_count} as an \code{offset}, \code{condition_group} as a predictor.
The \code{count} column is the response variable.

Since a \code{log} link is used, and \code{predicted_null_count} is estimated from native control batch effects, the extracted
coefficient estimates can be intreprested as log(fold change) realtive to the negative control.
}
\seealso{
\link{glm}
}

