% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concensus-objects.R
\name{newConcensusDataSet}
\alias{newConcensusDataSet}
\title{Set up analysis container for ConcensusGLM}
\usage{
newConcensusDataSet(checkpoint = FALSE, working_directory = ".",
  load_checkpoint = NULL, ...)
}
\arguments{
\item{checkpoint}{Logical. Save intermediate results as checkpoints?}

\item{working_directory}{Character. Path to where you want the analysis output.}

\item{load_checkpoint}{If \code{"latest"}, reload from latest checkpoint files in \code{working_directory}.
If a string pattern matching existing checkpoint, then loads those files. If \code{NULL} (the default) then doesn nothing.}
}
\value{
list of class "concensusDataSet"
}
\description{
Set up analysis container for ConcensusGLM. This is the workhorse function; most users should use
\link{concensusDataSetFromFile}.
}
\seealso{
\link{workflows::newWorkflow}, \link{concensusDataSetFromFile}
}

