% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concensus-methods.R
\name{print.concensusDataSet}
\alias{print.concensusDataSet}
\title{Print ConcensusDataSet}
\usage{
\method{print}{concensusDataSet}(x, ...)
}
\arguments{
\item{x}{ConcensusDataSet.}

\item{...}{Other arguments.}
}
\value{
\code{x}, invisibly.
}
\description{
Print summary stitistics of a ConcensusDataSet to the console
}

