% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concensus-methods.R
\name{print.concensusWorkflow}
\alias{print.concensusWorkflow}
\title{Print ConcensusWorkflow}
\usage{
\method{print}{concensusWorkflow}(x, ...)
}
\arguments{
\item{x}{ConcensusWorkflow.}

\item{...}{Other arguments.}
}
\description{
Print data of a ConcensusWorkflow to the console
}

