% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concensus-statistical-methods.R
\name{resample}
\alias{resample}
\alias{resample.concensusDataSet}
\alias{resample.concensusWorkflow}
\alias{resample.default}
\title{Make fake data from resampled controls}
\usage{
resample(x, ...)

\method{resample}{default}(x, ...)

\method{resample}{concensusWorkflow}(x, ...)

\method{resample}{concensusDataSet}(x, n_replicates = 2, n_samples = 10000,
  prevalence = 0.5, grouping = c("id", "plate_name", "well"),
  positive_control. = NULL, ...)
}
\arguments{
\item{x}{concensusWorkflow or concensusDataSet.}

\item{...}{Other arguments.}

\item{n_replicates}{Numeric.}

\item{n_samples}{Numeric.}

\item{prevalence}{Numeric. Positive controls (if present) are resampled at a frequency of \code{n_samples * 2 * prevalence}.}

\item{grouping}{Character vector. Grouping of wells.}

\item{positive_control}{Character. Pattern to assign positive controls on. Overrides any already-present positive control.}
}
\value{
concensusWorkflow or concensusDataSet.
}
\description{
Make fake data from resampled controls.
}
\details{
None.
}

