% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concensus-methods.R
\name{View_c}
\alias{View_c}
\title{Invoke a Data Viewer for concensus objects}
\usage{
View_c(x, ...)
}
\arguments{
\item{x}{concensusWorkflow}

\item{...}{Other arguments.}

\item{scatter_n}{Numeric. If \code{x} is a \code{concensusWorkflow} which has been scattered, use this chunk.}

\item{element}{Character. Which element of \code{concensusDataset} to look at. Defaults to the input data, \code{"data"}.}

\item{limit_size}{Numeric. Limit number of rows passed to \code{View}. For very large data, this can stop
your RStudio session from crashing! Default \code{1000}.}
}
\description{
Invoke a spreadsheet-style data viewer on a Concensus R object.
}
\seealso{
\link{View}
}

