% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concensus-methods.R
\name{write_concensusDataSet}
\alias{write_concensusDataSet}
\title{Write ConcensusDataSet to disk}
\usage{
write_concensusDataSet(x, filename, clean = TRUE, ...)
}
\arguments{
\item{x}{ConcensusDataSet.}

\item{filename}{Character. Where to save.}

\item{...}{Other arguments passed to \code{write_workflow}.}
}
\description{
Serialize ConcensusDataSet to disk under \code{filename}. A wrapper to \code{write_workflow} from the
\code{workflows} package.
}
\seealso{
\link{saveRDS}, \link{write_workflow}
}

