% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concensus-methods.R
\name{write_plaintext}
\alias{View_c.concensusDataSet}
\alias{View_c.concensusWorkflow}
\alias{View_c.default}
\alias{write_plaintext}
\alias{write_plaintext.concensusWorkflow}
\alias{write_plaintext.default}
\title{Write ConcensusGLM data as plain text}
\usage{
write_plaintext(x, ...)

\method{write_plaintext}{default}(x, ...)

\method{write_plaintext}{concensusWorkflow}(x, filename, input_csv = TRUE,
  output_csv = TRUE, output_matrix = TRUE, ...)

\method{View_c}{default}(x, ...)

\method{View_c}{concensusWorkflow}(x, scatter_n = 1, element = "data",
  limit_size = 1000, ...)

\method{View_c}{concensusDataSet}(x, element = "data", limit_size = 1000,
  ...)
}
\arguments{
\item{x}{concensusWorkflow}

\item{...}{Other arguments.}

\item{input_csv}{Character. Write.}

\item{output_csv}{Character. Where to save.}

\item{output_matrix}{Logical. Save a matrix of strains and conditions. Make sure concentrations (if used for
final fit) are common across all compounds or this will make a mess.}
}
\description{
Write ConcensusGLM data as plain text
}

