% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concensus-methods.R
\name{clean}
\alias{clean}
\alias{clean.data.frame}
\alias{clean.default}
\title{Clean up ConcensusGLM data}
\usage{
clean(x, ...)

\method{clean}{default}(x, ...)

\method{clean}{data.frame}(x, threshold = 100, plate_fraction = 0.6, ...)
}
\arguments{
\item{x}{data.frame to clean.}

\item{...}{Other arguments.}

\item{threshold}{Numeric. Strains below this total count threshold will be discarded.
Plates below 1000 x \code{threshold} will be discarded.}
}
\value{
Cleaned data.frame
}
\description{
Clean up ConcensusGLM data. Removes strains with total count across data below \code{threshold} and plates
with total counts below 1000 x \code{threshold}
}

