% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concensus-methods.R
\name{gather.concensusWorkflow}
\alias{gather.concensusWorkflow}
\title{Gather concensusWorkflow}
\usage{
\method{gather}{concensusWorkflow}(x, ...)
}
\arguments{
\item{x}{concensusWorkflow.}

\item{...}{Other arguments passed to \code{workflows::execute}.}
}
\value{
Gathered concensusWorkflow
}
\description{
Gather concensusWorkflow back into a single pipeline.
}
\seealso{
\link{gather}
}

