% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concensus-methods.R
\name{write_concensusDataSet}
\alias{write_concensusDataSet}
\title{Write ConcensusDataSet to disk}
\usage{
write_concensusDataSet(x, filename, ...)
}
\arguments{
\item{x}{ConcensusDataSet.}

\item{filename}{Character. Where to save.}

\item{...}{Other arguments.}
}
\description{
Serialize ConcensusDataSet to disk under \code{filename}.
}
\seealso{
\link{saveRDS}
}

